refactor(simulation): implement SoA architecture and index-based QuadTree
- Refactor math primitives (Vector2, Rectangle) to f32.
- Implement Simulation struct with Structure of Arrays (SoA) layout.
- Update QuadTree to store particle indices and work with SoA state.
- Expose SimulationWrapper via wasm_bindgen.

+++ nbody_simulator/src/rectangle.rs
@@ -3,12 +3,12 @@ use crate::vector2::Vector2;
 #[derive(Debug, Copy, Clone)]
 pub struct Rectangle {
     pub position: Vector2,
-    pub width: f64,
-    pub height: f64,
+    pub width: f32,
+    pub height: f32,
 }
 
 impl Rectangle {
-    pub fn new(position: Vector2, width: f64, height: f64) -> Rectangle {
+    pub fn new(position: Vector2, width: f32, height: f32) -> Rectangle {
         Rectangle {
             position,
             width,
